shared/py/requirements.txt
confluent_kafka==2.12.2
fastapi==0.121.1
httpx==0.28.1
motor==3.6.0
pathway==0.26.4
prisma==0.15.0
pydantic==2.12.4
pymongo==4.9.2
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.2.1
redis==5.2.1
Requests==2.32.5
uvicorn==0.38.0
websockets==15.0.1
